Move copy_dir (and test) into common/cgmcore/file_utils.py
