+++ src/common/evaluation/QA/eval_depthmap_models/src/test_evaluate.py
@@ -7,54 +7,12 @@ import pandas as pd
 
 CWD = Path(__file__).resolve()
 
-sys.path.append(str(CWD.parents[4]))
+sys.path.append(str(CWD.parents[4]))  # common/ dir
 
 REPO_DIR = str(CWD.parents[6].absolute())
 
-from evaluation.QA.eval_depthmap_models.src.evaluate import (copy_dir, prepare_sample_dataset,  # noqa: E402
+from evaluation.QA.eval_depthmap_models.src.evaluate import (prepare_sample_dataset,  # noqa: E402
                                                              tf_load_pickle)
-
-
-@pytest.fixture
-def temp_common_dir():
-    temp_common_dir = Path(CWD.parent / "temp_common")
-    yield temp_common_dir
-    try:
-        shutil.rmtree(temp_common_dir)
-    except OSError:
-        pass
-
-
-@pytest.fixture
-def empty_dir():
-    empty_dir = Path(CWD.parent / "copy_empty")
-    empty_dir.mkdir(parents=True, exist_ok=True)
-    yield empty_dir
-    try:
-        shutil.rmtree(empty_dir)
-    except OSError:
-        pass
-
-
-@pytest.fixture
-def temp_empty_dir():
-    temp_empty_dir = Path(CWD.parent / "temp_empty_dir")
-    yield temp_empty_dir
-    try:
-        shutil.rmtree(temp_empty_dir)
-    except OSError:
-        pass
-
-
-def test_copy_dir(temp_common_dir):
-    common_dir_path = Path(REPO_DIR + "/src/common")
-    copy_dir(src=common_dir_path, tgt=temp_common_dir, glob_pattern='*/*.py', should_touch_init=True)
-    assert temp_common_dir.is_dir(), 'The temp_common_dir does not exist. Did copy_dir fail?'
-
-
-def test_copy_empty_dir(empty_dir, temp_empty_dir):
-    copy_dir(src=empty_dir, tgt=temp_empty_dir, glob_pattern='*/*.py', should_touch_init=False)
-    assert temp_empty_dir.is_dir(), 'The temp_empty_dir does not exist. Did copy_dir fail for an empty directory?'
 
 
 def test_prepare_sample_dataset():
